Module_1/Week_4/object_dectection/requirements.txt
streamlit
pillow 
numpy
opencv-python